# backend/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.5"
  },
  "host": "gihitmxztbocbbdbgsqx.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/commands": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.commands.id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.node_id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.command"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.status"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.executed_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/commands"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "commands"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.commands"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "commands"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.commands.id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.node_id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.command"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.status"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.executed_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "commands"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.commands.id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.gateway_id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.node_id"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.command"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.status"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.commands.executed_at"
          },
          {
            "$ref": "#/parameters/body.commands"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "commands"
        ]
      }
    },
...
  },
  "definitions": {
...
    "commands": {
      "required": [
        "id",
        "command"
      ],
      "properties": {
        "id": {
          "default": "gen_random_uuid()",
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "uuid",
          "type": "string"
        },
        "gateway_id": {
          "description": "Note:\nThis is a Foreign Key to `gateways.id`.<fk table='gateways' column='id'/>",
          "format": "uuid",
          "type": "string"
        },
        "node_id": {
          "description": "Note:\nThis is a Foreign Key to `nodes.id`.<fk table='nodes' column='id'/>",
          "format": "uuid",
          "type": "string"
        },
        "command": {
          "format": "text",
          "type": "string"
        },
        "status": {
          "default": "pending",
          "format": "text",
          "type": "string"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "executed_at": {
          "format": "timestamp without time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.commands": {
      "name": "commands",
      "description": "commands",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/commands"
      }
    },
    "rowFilter.commands.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.commands.gateway_id": {
      "name": "gateway_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.commands.node_id": {
      "name": "node_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.commands.command": {
      "name": "command",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.commands.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.commands.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.commands.executed_at": {
      "name": "executed_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}